refactor; replacing information rules
